Add an extra Canvas UI element for scrolling background
I did this because there was an unexpected bug introduced to the UI when I switched from Overlay to Camera space with the Canvas object containing the ability buttons (e.g. shield and laser). My solution was to use 2 separate canvas UI objects to hold an overlay for the abilities and another to hold the background scrolling effect