Fix broken reference to energy absorber script
I added a reference to the shield gameObject in the PlayerLaser script to resolve nullreference exception with Shield gameobject.Do note that you cannot find an inactive gameobject in the scene using GameObject.Find(). The game's controls can be found in TouchControlsFinal scene. If you are going to change anything in the scene, please make a copy of it as merge conflicts can occur with unity scene files when multiple people are working on the same scene. Shield needs to be toggled on/off and requires a shield energy bar that slowly regenerates over time when disabled, and depletes when shield is active. Let me handle this part of the code and everyone else needs to focus on creating/designing bullet patterns and movement patterns for the enemies